Refactor test_jit and separate model testing
