+++ test/model_defs/__init__.py
@@ -1,0 +1,6 @@
+from .alexnet import *
+from .vgg import *
+from .resnet import *
+from .densenet import *
+from .squeezenet import *
+from .inception import *
